requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
Django==5.2.6
django-cors-headers==4.8.0
django-crispy-forms==2.4
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
